--- fe-house-p/src/components/MovilNavBar/index.js
@@ -1,3 +1,0 @@
-import MovilNavbar from "./MovilNavBar";
-
-export { MovilNavbar }
