+++ fe-house-p/src/components/Navbar/index.js
@@ -1,0 +1,4 @@
+import DesktopNav from "./DesktopNav";
+import MovilNav from "./MovilNav";
+
+export {DesktopNav, MovilNav}
